MCQ Generator/requirements.txt
Flask
pdfplumber
python-docx
fpdf
google-generativeai
